Relax release-version-link rule
Which linted that a release heading must link to:
https://github.com/OWNER/REPO/compare/A...B

It now only lints that the heading has a link (to any URL).

The compare URL was tricky to get right in scenarios with multiple
release lines or hotfixes managed from git branches (e.g. v1.x or
v2.1.6) where the changelog entries do not align with git tags.

Relaxing this rule also means `remark-changelog` can lint
shallowly cloned repositories without tags (the default in GitHub
Actions). Previously it was necessary to run `git fetch --tags`
beforehand which doesn't work for GHA workflows triggered by a tag
on push.

In addition, fix mode now adds a link on the first release,
linking to a GitHub release. Releases that already have a link
are no longer touched.